Refactored name fix
